Integrate prepare-slurm into vara-run
